Lagged macro-vars & tROC on credit dataset
The portion that creates lagged macroeconomic variables was missing during my re-organizing of the codebase. It is now created as it should be in script 3c.

Apart from some general housekeeping, I also moved the analytical functions defined in script 5a(i) to their rightful place for such functional definitions in script 0b(i). Remember this practice in future!

In script 5a(i), and in fitting the final Cox model for the TFD-definition, I had to deselect the [pmnt_method_grp] variable, since it led to model failure; something to investigate when we refine this after the close-out. Regardless, this was done to facilitate time-dependent ROC-analyses using our new tROC() and tROC.multi() functions. I did a quick comparative study to show the runtime difference between either calls, though the idea is to use the multi-threaded version from now one. As such, we can keep the usual tROC()-call as comments, but only within this script 5a(i). Not necessary to show the same for the other time definitions.